refactor(manifest): extract comparison engine and analyzers (#601 PR3)
Extract three internal static classes from ProductionManifestComparer:

- ComparisonEngine: FindDuplicates + PerformComparison (pure prior-vs-new
  matching by Bates then Control Number, classifying added/removed/
  unchanged/changed/replaced/duplicated and populating summary counts).
- BatesRangeAnalyzer: AnalyzeBatesRanges + ResultSummaryBatesAnalysis +
  FindGapsInSequence + GroupBatesByPrefix + MergeConsecutiveBates +
  TryParseBates + FormatBates (pure Bates Number gap/overlap analysis).
- VolumeAnalyzer: AnalyzeVolumes (per-volume Bates range and status
  rollups).

ProductionManifestComparer.CompareAndReportAsync now delegates to these
classes. No behavior, JSON contract, or public surface change; the façade
signature is unchanged. Output parity verified: ComparisonTests and
ProductionManifestComparerTests still pass 29/29 through the façade.

Add direct unit tests for the now-internal pure types (BatesRangeAnalyzer
gap/parse/format and ComparisonEngine matching/duplicates) to cover the
extracted logic at the unit level (11 new tests).

Module file sizes: ProductionManifestComparer 402 -> comparison engine
219 + Bates analyzer 213 + volume analyzer 110. The remaining façade
holds the loader and report writers, which subsequent PRs will extract.

Third slice of the #601 god-class decomposition.
