+++ src/ManifestComparison/ProductionManifestComparer.cs
@@ -77,17 +77,17 @@ public static class ProductionManifestComparer
             }
             priorRecords.AddRange(recs);
         }
-        var priorDuplicates = FindDuplicates(priorRecords, "prior");
+        var priorDuplicates = ComparisonEngine.FindDuplicates(priorRecords, "prior");
 
         var newRecords = await LoadRecordsAsync(newManifest.ResolvedPath, newManifest.Manifest).ConfigureAwait(false);
         foreach (var r in newRecords)
         {
             r.ProductionId = newManifest.Manifest.ProductionId;
         }
-        var newDuplicates = FindDuplicates(newRecords, "new");
+        var newDuplicates = ComparisonEngine.FindDuplicates(newRecords, "new");
 
         // Normalize comparison and run matching logic
-        var result = PerformComparison(priorRecords, newRecords, mode, priorDuplicates, newDuplicates, priorManifests.Select(pm => pm.ResolvedPath).ToList(), newManifest.ResolvedPath);
+        var result = ComparisonEngine.PerformComparison(priorRecords, newRecords, mode, priorDuplicates, newDuplicates, priorManifests.Select(pm => pm.ResolvedPath).ToList(), newManifest.ResolvedPath);
 
         // Populate ranges
         result.BatesAnalysis.PriorRange = priorRecords.Count > 0
@@ -107,10 +107,10 @@ public static class ProductionManifestComparer
         }
 
         // Perform Bates range analysis and gap/overlap detection
-        AnalyzeBatesRanges(priorRecords, newRecords, mode, result.BatesAnalysis, result.Details);
+        BatesRangeAnalyzer.AnalyzeBatesRanges(priorRecords, newRecords, mode, result.BatesAnalysis, result.Details);
 
         // Perform Volume analysis
-        AnalyzeVolumes(priorRecords, newRecords, result.VolumeAnalysis);
+        VolumeAnalyzer.AnalyzeVolumes(priorRecords, newRecords, result.VolumeAnalysis);
 
         // Write report
         var jsonReport = JsonSerializer.Serialize(result, SerializerOptions);
@@ -298,520 +298,6 @@ public static class ProductionManifestComparer
         }
         fields.Add(currentField.ToString());
         return fields;
-    }
-
-    private static List<DuplicateDetail> FindDuplicates(List<ComparisonRecord> records, string set)
-    {
-        var duplicates = new List<DuplicateDetail>();
-        var batesSeen = new Dictionary<string, ComparisonRecord>(StringComparer.OrdinalIgnoreCase);
-        var controlSeen = new Dictionary<string, ComparisonRecord>(StringComparer.OrdinalIgnoreCase);
-
-        foreach (var r in records)
-        {
-            if (!string.IsNullOrEmpty(r.BatesNumber))
-            {
-                if (batesSeen.TryGetValue(r.BatesNumber, out var first))
-                {
-                    duplicates.Add(new DuplicateDetail
-                    {
-                        Set = set,
-                        BatesNumber = r.BatesNumber,
-                        ControlNumber = r.ControlNumber,
-                        Message = $"Duplicate Bates Number '{r.BatesNumber}' on line {r.SourceLine} (first seen on line {first.SourceLine})"
-                    });
-                }
-                else
-                {
-                    batesSeen[r.BatesNumber] = r;
-                }
-            }
-
-            if (!string.IsNullOrEmpty(r.ControlNumber))
-            {
-                if (controlSeen.TryGetValue(r.ControlNumber, out var first))
-                {
-                    duplicates.Add(new DuplicateDetail
-                    {
-                        Set = set,
-                        BatesNumber = r.BatesNumber,
-                        ControlNumber = r.ControlNumber,
-                        Message = $"Duplicate Control Number '{r.ControlNumber}' on line {r.SourceLine} (first seen on line {first.SourceLine})"
-                    });
-                }
-                else
-                {
-                    controlSeen[r.ControlNumber] = r;
-                }
-            }
-        }
-
-        return duplicates;
-    }
-
-    private static ComparisonResult PerformComparison(
-        List<ComparisonRecord> priorRecords,
-        List<ComparisonRecord> newRecords,
-        string mode,
-        List<DuplicateDetail> priorDuplicates,
-        List<DuplicateDetail> newDuplicates,
-        List<string> priorPaths,
-        string newPath)
-    {
-        var result = new ComparisonResult
-        {
-            ComparisonMode = mode,
-            Manifests = priorPaths.Concat(new[] { newPath }).ToList()
-        };
-
-        result.Details.Duplicates.AddRange(priorDuplicates);
-        result.Details.Duplicates.AddRange(newDuplicates);
-
-        var priorByBates = priorRecords.Where(r => !string.IsNullOrEmpty(r.BatesNumber))
-            .ToLookup(r => r.BatesNumber, StringComparer.OrdinalIgnoreCase);
-        var priorByControl = priorRecords.Where(r => !string.IsNullOrEmpty(r.ControlNumber))
-            .ToLookup(r => r.ControlNumber, StringComparer.OrdinalIgnoreCase);
-
-        var matchedPriorBates = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        var matchedPriorControls = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-
-        foreach (var nr in newRecords)
-        {
-            bool isOverlap = false;
-            if (string.Equals(mode, "supplemental", StringComparison.OrdinalIgnoreCase))
-            {
-                var priorBatesOverlap = priorByBates.Contains(nr.BatesNumber);
-                var priorControlOverlap = priorByControl.Contains(nr.ControlNumber);
-
-                if (priorBatesOverlap || priorControlOverlap)
-                {
-                    result.Details.Duplicates.Add(new DuplicateDetail
-                    {
-                        Set = "new",
-                        BatesNumber = nr.BatesNumber,
-                        ControlNumber = nr.ControlNumber,
-                        Message = $"Supplemental overlap: Bates '{nr.BatesNumber}' or Control '{nr.ControlNumber}' matches a prior production set."
-                    });
-                    isOverlap = true;
-                }
-            }
-
-            if (isOverlap)
-            {
-                result.Details.Added.Add(new RecordDetail
-                {
-                    BatesNumber = nr.BatesNumber,
-                    ControlNumber = nr.ControlNumber,
-                    FilePath = nr.FilePath,
-                    Hash = nr.Hash,
-                    Volume = nr.Volume
-                });
-                continue;
-            }
-
-            // 1. Try to match by Bates Number
-            var priorBatesMatch = priorByBates[nr.BatesNumber].FirstOrDefault();
-            if (priorBatesMatch is not null)
-            {
-                matchedPriorBates.Add(priorBatesMatch.BatesNumber);
-                matchedPriorControls.Add(priorBatesMatch.ControlNumber);
-
-                bool isHashMatch = true;
-                if (!string.IsNullOrEmpty(nr.Hash) && !string.IsNullOrEmpty(priorBatesMatch.Hash))
-                {
-                    isHashMatch = string.Equals(nr.Hash, priorBatesMatch.Hash, StringComparison.OrdinalIgnoreCase);
-                }
-
-                bool isPathMatch = string.Equals(nr.FilePath, priorBatesMatch.FilePath, StringComparison.OrdinalIgnoreCase);
-                bool isControlMatch = string.Equals(nr.ControlNumber, priorBatesMatch.ControlNumber, StringComparison.OrdinalIgnoreCase);
-
-                if (isHashMatch && isPathMatch && isControlMatch)
-                {
-                    result.Details.Unchanged.Add(new RecordDetail
-                    {
-                        BatesNumber = nr.BatesNumber,
-                        ControlNumber = nr.ControlNumber,
-                        FilePath = nr.FilePath,
-                        Hash = nr.Hash,
-                        Volume = nr.Volume
-                    });
-                }
-                else
-                {
-                    result.Details.Changed.Add(new ChangedDetail
-                    {
-                        BatesNumber = nr.BatesNumber,
-                        ControlNumber = nr.ControlNumber,
-                        PriorPath = priorBatesMatch.FilePath,
-                        NewPath = nr.FilePath,
-                        PriorHash = priorBatesMatch.Hash,
-                        NewHash = nr.Hash
-                    });
-                }
-                continue;
-            }
-
-            // 2. Try to match by Control Number
-            var priorControlMatch = priorByControl[nr.ControlNumber].FirstOrDefault();
-            if (priorControlMatch is not null)
-            {
-                matchedPriorBates.Add(priorControlMatch.BatesNumber);
-                matchedPriorControls.Add(priorControlMatch.ControlNumber);
-
-                result.Details.Replaced.Add(new ReplacedDetail
-                {
-                    PriorBatesNumber = priorControlMatch.BatesNumber,
-                    NewBatesNumber = nr.BatesNumber,
-                    ControlNumber = nr.ControlNumber
-                });
-                continue;
-            }
-
-            // 3. No match -> Added
-            result.Details.Added.Add(new RecordDetail
-            {
-                BatesNumber = nr.BatesNumber,
-                ControlNumber = nr.ControlNumber,
-                FilePath = nr.FilePath,
-                Hash = nr.Hash,
-                Volume = nr.Volume
-            });
-        }
-
-        // A supplemental manifest is additive and cannot establish removals.
-        if (!string.Equals(mode, "supplemental", StringComparison.OrdinalIgnoreCase))
-        {
-            foreach (var pr in priorRecords)
-            {
-                if (!matchedPriorBates.Contains(pr.BatesNumber) && !matchedPriorControls.Contains(pr.ControlNumber))
-                {
-                    result.Details.Removed.Add(new RecordDetail
-                    {
-                        BatesNumber = pr.BatesNumber,
-                        ControlNumber = pr.ControlNumber,
-                        FilePath = pr.FilePath,
-                        Hash = pr.Hash,
-                        Volume = pr.Volume
-                    });
-                }
-            }
-        }
-
-        // Populate summary counts
-        result.Summary.TotalPriorRecords = priorRecords.Count;
-        result.Summary.TotalNewRecords = newRecords.Count;
-        result.Summary.AddedCount = result.Details.Added.Count;
-        result.Summary.RemovedCount = result.Details.Removed.Count;
-        result.Summary.UnchangedCount = result.Details.Unchanged.Count;
-        result.Summary.ChangedCount = result.Details.Changed.Count;
-        result.Summary.ReplacedCount = result.Details.Replaced.Count;
-        result.Summary.DuplicateCount = result.Details.Duplicates.Count;
-
-        return result;
-    }
-
-    private static void AnalyzeBatesRanges(
-        List<ComparisonRecord> priorRecords,
-        List<ComparisonRecord> newRecords,
-        string mode,
-        BatesAnalysis batesAnalysis,
-        ResultDetails details)
-    {
-        // 1. Group records by Bates Prefix
-        var priorBates = priorRecords.Select(r => r.BatesNumber).Where(b => !string.IsNullOrEmpty(b)).ToList();
-        var newBates = newRecords.Select(r => r.BatesNumber).Where(b => !string.IsNullOrEmpty(b)).ToList();
-
-        // Skipped Bates Numbers / Gap detection within the new set
-        var gaps = FindGapsInSequence(newBates);
-        details.Skipped.AddRange(gaps.Select(g => new BatesRangeReport { Start = g.Start, End = g.End }));
-        batesAnalysis.Gaps.AddRange(details.Skipped);
-
-        // Gap detection between prior max and new min in supplemental mode
-        if (string.Equals(mode, "supplemental", StringComparison.OrdinalIgnoreCase) && priorBates.Count > 0 && newBates.Count > 0)
-        {
-            // Group by prefix to find boundaries
-            var priorGroups = GroupBatesByPrefix(priorBates);
-            var newGroups = GroupBatesByPrefix(newBates);
-
-            foreach (var prefix in newGroups.Keys)
-            {
-                if (priorGroups.TryGetValue(prefix, out var priorGroup))
-                {
-                    var maxPrior = priorGroup.Max(g => g.Value);
-                    var minNew = newGroups[prefix].Min(g => g.Value);
-                    var digits = newGroups[prefix].First().Digits;
-
-                    if (minNew > maxPrior + 1)
-                    {
-                        var gapStart = FormatBates(maxPrior + 1, prefix, digits);
-                        var gapEnd = FormatBates(minNew - 1, prefix, digits);
-                        batesAnalysis.Gaps.Add(new BatesRangeReport { Start = gapStart, End = gapEnd });
-                    }
-                }
-            }
-        }
-
-        // Bates Number overlaps (duplicates between sets)
-        if (priorBates.Count > 0 && newBates.Count > 0)
-        {
-            var priorSet = new HashSet<string>(priorBates, StringComparer.OrdinalIgnoreCase);
-            var overlaps = new List<string>();
-            foreach (var b in newBates)
-            {
-                if (priorSet.Contains(b))
-                {
-                    overlaps.Add(b);
-                }
-            }
-
-            if (overlaps.Count > 0)
-            {
-                var mergedOverlaps = MergeConsecutiveBates(overlaps);
-                batesAnalysis.Overlaps.AddRange(mergedOverlaps);
-            }
-        }
-
-        ResultSummaryBatesAnalysis(batesAnalysis);
-    }
-
-    private static void ResultSummaryBatesAnalysis(BatesAnalysis batesAnalysis)
-    {
-        // Update summary skipped count
-        int totalSkipped = 0;
-        foreach (var gap in batesAnalysis.Gaps)
-        {
-            if (TryParseBates(gap.Start, out var p1, out long v1, out _) && TryParseBates(gap.End, out var p2, out long v2, out _) && p1 == p2)
-            {
-                totalSkipped += (int)(v2 - v1 + 1);
-            }
-            else
-            {
-                totalSkipped++;
-            }
-        }
-        batesAnalysis.TotalSkippedBates = totalSkipped;
-    }
-
-    private static List<BatesRangeReport> FindGapsInSequence(List<string> batesList)
-    {
-        var gaps = new List<BatesRangeReport>();
-        if (batesList.Count == 0) return gaps;
-
-        var parsed = GroupBatesByPrefix(batesList);
-
-        foreach (var kvp in parsed)
-        {
-            var prefix = kvp.Key;
-            var items = kvp.Value.OrderBy(i => i.Value).ToList();
-            if (items.Count <= 1) continue;
-
-            for (int i = 0; i < items.Count - 1; i++)
-            {
-                var current = items[i].Value;
-                var next = items[i + 1].Value;
-                var digits = items[i].Digits;
-
-                if (next > current + 1)
-                {
-                    gaps.Add(new BatesRangeReport
-                    {
-                        Start = FormatBates(current + 1, prefix, digits),
-                        End = FormatBates(next - 1, prefix, digits)
-                    });
-                }
-            }
-        }
-
-        return gaps;
-    }
-
-    private static Dictionary<string, List<(long Value, int Digits)>> GroupBatesByPrefix(List<string> batesList)
-    {
-        var groups = new Dictionary<string, List<(long Value, int Digits)>>(StringComparer.OrdinalIgnoreCase);
-        foreach (var b in batesList)
-        {
-            if (TryParseBates(b, out var prefix, out long val, out int digits))
-            {
-                if (!groups.TryGetValue(prefix, out var list))
-                {
-                    list = new List<(long Value, int Digits)>();
-                    groups[prefix] = list;
-                }
-                list.Add((val, digits));
-            }
-        }
-        return groups;
-    }
-
-    private static List<BatesRangeReport> MergeConsecutiveBates(List<string> batesList)
-    {
-        var merged = new List<BatesRangeReport>();
-        if (batesList.Count == 0) return merged;
-
-        var parsed = GroupBatesByPrefix(batesList);
-
-        foreach (var kvp in parsed)
-        {
-            var prefix = kvp.Key;
-            var items = kvp.Value.OrderBy(i => i.Value).Select(i => i.Value).Distinct().ToList();
-            if (items.Count == 0) continue;
-
-            long start = items[0];
-            long last = items[0];
-            int digits = kvp.Value[0].Digits;
-
-            for (int i = 1; i < items.Count; i++)
-            {
-                if (items[i] == last + 1)
-                {
-                    last = items[i];
-                }
-                else
-                {
-                    merged.Add(new BatesRangeReport
-                    {
-                        Start = FormatBates(start, prefix, digits),
-                        End = FormatBates(last, prefix, digits)
-                    });
-                    start = items[i];
-                    last = items[i];
-                }
-            }
-
-            merged.Add(new BatesRangeReport
-            {
-                Start = FormatBates(start, prefix, digits),
-                End = FormatBates(last, prefix, digits)
-            });
-        }
-
-        return merged;
-    }
-
-    private static bool TryParseBates(string bates, out string prefix, out long num, out int digits)
-    {
-        prefix = string.Empty;
-        num = 0;
-        digits = 0;
-        if (string.IsNullOrEmpty(bates)) return false;
-
-        int idx = bates.Length - 1;
-        while (idx >= 0 && char.IsDigit(bates[idx]))
-        {
-            idx--;
-        }
-
-        prefix = bates.Substring(0, idx + 1);
-        string numPart = bates.Substring(idx + 1);
-        if (numPart.Length == 0) return false;
-
-        digits = numPart.Length;
-        return long.TryParse(numPart, System.Globalization.CultureInfo.InvariantCulture, out num);
-    }
-
-    private static string FormatBates(long value, string prefix, int digits)
-    {
-        return $"{prefix}{value.ToString($"D{digits}", System.Globalization.CultureInfo.InvariantCulture)}";
-    }
-
-    private static void AnalyzeVolumes(
-        List<ComparisonRecord> priorRecords,
-        List<ComparisonRecord> newRecords,
-        List<VolumeResult> volumeAnalysis)
-    {
-        var priorGroups = priorRecords.GroupBy(r => r.ProductionId, StringComparer.OrdinalIgnoreCase)
-            .ToDictionary(g => g.Key, g => g.ToList(), StringComparer.OrdinalIgnoreCase);
-
-        var newVols = newRecords.GroupBy(r => r.Volume, StringComparer.OrdinalIgnoreCase)
-            .ToDictionary(g => g.Key, g => g.ToList(), StringComparer.OrdinalIgnoreCase);
-
-        foreach (var priorKvp in priorGroups)
-        {
-            var priorProdId = priorKvp.Key;
-            var priorRecs = priorKvp.Value;
-
-            var priorVols = priorRecs.GroupBy(r => r.Volume, StringComparer.OrdinalIgnoreCase)
-                .ToDictionary(g => g.Key, g => g.ToList(), StringComparer.OrdinalIgnoreCase);
-
-            var volNames = priorVols.Keys.OrderBy(v => v).ToList();
-
-            foreach (var vol in volNames)
-            {
-                var priorList = priorVols[vol];
-                bool inNew = newVols.TryGetValue(vol, out var newList);
-
-                var priorRange = priorList.Count > 0
-                    ? $"{priorList.Min(r => r.BatesNumber)} - {priorList.Max(r => r.BatesNumber)}"
-                    : string.Empty;
-
-                var newRange = inNew && newList is not null && newList.Count > 0
-                    ? $"{newList.Min(r => r.BatesNumber)} - {newList.Max(r => r.BatesNumber)}"
-                    : string.Empty;
-
-                var status = "unchanged";
-                if (!inNew)
-                {
-                    status = "removed";
-                }
-                else if (newList is not null)
-                {
-                    var priorSet = new Dictionary<string, ComparisonRecord>(StringComparer.OrdinalIgnoreCase);
-                    foreach (var r in priorList)
-                    {
-                        if (!string.IsNullOrEmpty(r.BatesNumber))
-                        {
-                            priorSet[r.BatesNumber] = r;
-                        }
-                    }
-
-                    bool hasChanges = false;
-                    foreach (var nr in newList)
-                    {
-                        if (priorSet.TryGetValue(nr.BatesNumber, out var pr))
-                        {
-                            if (!string.Equals(nr.FilePath, pr.FilePath, StringComparison.OrdinalIgnoreCase) ||
-                                (!string.IsNullOrEmpty(nr.Hash) && !string.IsNullOrEmpty(pr.Hash) && !string.Equals(nr.Hash, pr.Hash, StringComparison.OrdinalIgnoreCase)))
-                            {
-                                hasChanges = true;
-                                break;
-                            }
-                        }
-                    }
-
-                    if (hasChanges || priorList.Count != newList.Count)
-                    {
-                        status = "changed";
-                    }
-                }
-
-                volumeAnalysis.Add(new VolumeResult
-                {
-                    ProductionId = priorProdId,
-                    VolumeName = vol,
-                    PriorBatesRange = priorRange,
-                    NewBatesRange = newRange,
-                    Status = status
-                });
-            }
-        }
-
-        // Also identify volumes in the new set that do not exist in ANY prior set
-        var newVolsOnly = newVols.Keys.Where(v => !priorRecords.Any(r => string.Equals(r.Volume, v, StringComparison.OrdinalIgnoreCase))).OrderBy(v => v).ToList();
-        foreach (var vol in newVolsOnly)
-        {
-            var newList = newVols[vol];
-            var newRange = newList.Count > 0
-                ? $"{newList.Min(r => r.BatesNumber)} - {newList.Max(r => r.BatesNumber)}"
-                : string.Empty;
-
-            volumeAnalysis.Add(new VolumeResult
-            {
-                ProductionId = newRecords.FirstOrDefault()?.ProductionId ?? "NewSet",
-                VolumeName = vol,
-                PriorBatesRange = string.Empty,
-                NewBatesRange = newRange,
-                Status = "added"
-            });
-        }
     }
 
     private static string GenerateMarkdownSummary(ComparisonResult result, string mode)
